backend/requirements.txt
# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0

# Auth
bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0

# Environment
python-dotenv>=1.0.0

# APIs
openai>=1.12.0
google-genai>=0.3.0

# Audio processing (system: ffmpeg required)
